Remove UserController & move action in right controller (settingsController and userManagementController)
