+++ app/Resources/views/Admin/overview.html.twig
@@ -38,6 +38,7 @@
                             <td>
                                 <a href="#">View</a>
                                 <a href="#">Edit</a>
+                                <a href="#">Archive</a>
                                 <a href="#">Delete</a>
                             </td>
                         </tr>
