+++ src/GyverBundle/Controller/Settings/SettingsController.php
@@ -31,4 +31,17 @@ class SettingsController extends Controller
         return $this->render('Settings/home.html.twig');
     }
 
+    /**
+     * @Route("/account", name="user_info")
+     * @Method("GET")
+     * @return \Symfony\Component\HttpFoundation\Response
+     */
+    public function readAction()
+    {
+        //To get the user who is logged
+        $user = $this->get('security.token_storage')->getToken()->getUser();
+
+        return $this->render(':User:form_informations.html.twig', array('user' => $user ));
+    }
+
 }
