--- src/GyverBundle/Resources/views/User/showUser.html.twig
@@ -1,22 +1,0 @@
-{% extends "::base.html.twig" %}
-
-{% block title %}GyverBundle:User:showUser{% endblock %}
-
-{% block body %}
-    <h1>Welcome to the User:showUser page</h1>
-
-    <h2>{{ user.civility ? 'Monsieur' : 'Madame' }} {{ user.firstName}} {{ user.lastName}}</h2>
-
-    {% if user.company|length == 0 %}
-        no company
-    {% endif %}
-    {% for company in user.company %}
-        <li>{{ company.name }}</li>
-    {% endfor %}
-    {% for email in user.emailList %}
-        <li>{{ email.email }}</li>
-    {% endfor %}
-    {% for phoneNumber in user.phoneNumber %}
-        <li>{{ phoneNumber.phoneNumber }}</li>
-    {% endfor %}
-{% endblock %}
